(Debug): Organizando funções e melhorando fluxo entre services e controller
